testtask/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
PyMySQL
requests
gunicorn



